added benchmarks
