+++ benchmarks/series.cpp
@@ -39,5 +39,18 @@ int main(int argc, char *argv[])
                      .count()
               << "ms" << std::endl;
 
+
+    auto arg = add(x, pow(x, integer(2)));
+    auto ex = mul(sin(arg), cos(arg));
+    std::cout << "Expanding: " << *ex << std::endl;
+
+    t1 = std::chrono::high_resolution_clock::now();
+    auto res = SymEngine::UnivariateSeries::series(ex, "x", N);
+    t2 = std::chrono::high_resolution_clock::now();
+    // std::cout << *res[N-1] << std::endl;
+    std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t2 - t1)
+                     .count()
+              << "ms" << std::endl;
+
     return 0;
 }
