+++ symengine/polynomial.h
@@ -389,429 +389,6 @@ public:
     }
 }; // UnivariateExprPolynomial
 
-// MultivariatePolynomials
-class MultivariateIntPolynomial : public Basic
-{
-public:
-    // vars: set of variables for the polynomial
-    // degrees: max degrees of the symbols
-    // dict: dictionary for sparse represntation of polynomial, x**1 * y**2 + 3
-    // * x**4 * y ** 5
-    // is represented as {(1,2):1,(4,5):3}
-    set_sym vars_;
-    umap_sym_uint degrees_;
-    umap_uvec_mpz dict_;
-
-public:
-    IMPLEMENT_TYPEID(MULTIVARIATEINTPOLYNOMIAL)
-    // constructor from components
-    MultivariateIntPolynomial(const set_sym &var, umap_sym_uint &degrees,
-                              umap_uvec_mpz &dict);
-    // creates a MultivariateIntPolynomial in cannonical form based on
-    // dictionary d.
-    static RCP<const MultivariateIntPolynomial> from_dict(const set_sym &s,
-                                                          umap_uvec_mpz &&d);
-    vec_basic get_args() const;
-    bool is_canonical(const set_sym &vars, const umap_sym_uint &degrees,
-                      const umap_uvec_mpz &dict);
-    std::size_t __hash__() const;
-    bool __eq__(const Basic &o) const;
-    int compare(const Basic &o) const;
-    integer_class eval(std::map<RCP<const Symbol>, integer_class,
-                                RCPSymbolCompare> &vals) const;
-};
-
-// reconciles the positioning of the exponents in the vectors in the
-// umap_uvec_mpz dict_ of the arguments
-// with the positioning of the exponents in the correspondng vectors of the
-// output of the function.
-// v1 and v2 are vectors whose indices are the positions in the arguments and
-// whose values are the
-// positions in the output.  set_sym s is the set of symbols of the output, and
-// s1 and s2 are the sets of the symbols of the inputs.
-unsigned int reconcile(vec_uint &v1, vec_uint &v2, set_sym &s,
-                       const set_sym &s1, const set_sym &s2);
-// same as above, but for reconciling representation of a UnivariatePolynomial.
-unsigned int reconcile(vec_uint &v1, unsigned int &v2, set_sym &s,
-                       const set_sym &s1, const RCP<const Symbol> s2);
-// translates vectors of exponents from one polynomial into vectors of exponents
-// for another.
-// i.e. a polynomial if one of the addends is a polynomial in x,y and the sum is
-// a polynomial
-// in x,y,z, we will need to translate vectors with two elements to vectors with
-// three elements.
-// This situation can arise when we want to preform operations with two
-// polynomials in different
-// variables
-// translator is the vector generated by the reconcile function and size is the
-// size of the
-// resulting vector.
-vec_uint translate(vec_uint original, vec_uint translator, unsigned int size);
-// translates terms of UnivariateIntPolynomial into vectors
-vec_uint translate(unsigned int original, unsigned int translator,
-                   unsigned int size);
-vec_uint uint_vec_translate_and_add(const vec_uint &v1, const vec_uint &v2,
-                                    const vec_uint &translator1,
-                                    const vec_uint &translator2,
-                                    const unsigned int size);
-vec_uint uint_vec_translate_and_add(const vec_uint &v1, const unsigned int v2,
-                                    const vec_uint &translator1,
-                                    const unsigned int &translator2,
-                                    const unsigned int size);
-
-unsigned int reconcile(vec_int &v1, vec_int &v2, set_sym &s, const set_sym &s1,
-                       const set_sym &s2);
-unsigned int reconcile(vec_int &v1, unsigned int &v2, set_sym &s,
-                       const set_sym &s1, const RCP<const Symbol> s2);
-
-vec_int translate(vec_int original, vec_uint translator, unsigned int size);
-vec_int translate(int original, unsigned int translator, unsigned int size);
-vec_int int_vec_translate_and_add(const vec_int &v1, const vec_int &v2,
-                                  const vec_uint &translator1,
-                                  const vec_uint &translator2,
-                                  const unsigned int size);
-vec_int int_vec_translate_and_add(const vec_int &v1, const unsigned int v2,
-                                  const vec_uint &translator1,
-                                  const unsigned int &translator2,
-                                  const unsigned int size);
-
-RCP<const MultivariateIntPolynomial>
-add_mult_poly(const MultivariateIntPolynomial &a,
-              const MultivariateIntPolynomial &b);
-RCP<const MultivariateIntPolynomial>
-neg_mult_poly(const MultivariateIntPolynomial &a);
-RCP<const MultivariateIntPolynomial>
-sub_mult_poly(const MultivariateIntPolynomial &a,
-              const MultivariateIntPolynomial &b);
-RCP<const MultivariateIntPolynomial>
-mul_mult_poly(const MultivariateIntPolynomial &a,
-              const MultivariateIntPolynomial &b);
-
-RCP<const MultivariateIntPolynomial>
-add_mult_poly(const MultivariateIntPolynomial &a,
-              const UnivariateIntPolynomial &b);
-RCP<const MultivariateIntPolynomial>
-add_mult_poly(const UnivariateIntPolynomial &a,
-              const MultivariateIntPolynomial &b);
-RCP<const MultivariateIntPolynomial>
-sub_mult_poly(const MultivariateIntPolynomial &a,
-              const UnivariateIntPolynomial &b);
-RCP<const MultivariateIntPolynomial>
-sub_mult_poly(const UnivariateIntPolynomial &a,
-              const MultivariateIntPolynomial &b);
-RCP<const MultivariateIntPolynomial>
-mul_mult_poly(const MultivariateIntPolynomial &a,
-              const UnivariateIntPolynomial &b);
-RCP<const MultivariateIntPolynomial>
-mul_mult_poly(const UnivariateIntPolynomial &a,
-              const MultivariateIntPolynomial &b);
-
-RCP<const MultivariateIntPolynomial>
-add_mult_poly(const UnivariateIntPolynomial &a,
-              const UnivariateIntPolynomial &b);
-RCP<const MultivariateIntPolynomial>
-sub_mult_poly(const UnivariateIntPolynomial &a,
-              const UnivariateIntPolynomial &b);
-RCP<const MultivariateIntPolynomial>
-mul_mult_poly(const UnivariateIntPolynomial &a,
-              const UnivariateIntPolynomial &b);
-
-// MultivariatePolynomial
-class MultivariatePolynomial : public Basic
-{
-public:
-    // vars: set of variables for the polynomial
-    // degrees: max degrees of the symbols
-    // dict: dictionary for sparse represntation of polynomial, x**1 * y**2 + 3
-    // * x**4 * y ** 5
-    // is represented as {(1,2):1,(4,5):3}
-    set_sym vars_;
-    umap_sym_int degrees_;
-    umap_vec_expr dict_;
-
-public:
-    IMPLEMENT_TYPEID(MULTIVARIATEPOLYNOMIAL);
-    // constructor from components
-    MultivariatePolynomial(const set_sym &var, umap_sym_int &degrees,
-                           umap_vec_expr &dict);
-    // creates a MultivariatePolynomial in cannonical form based on dictionary
-    // d.
-    static RCP<const MultivariatePolynomial> from_dict(const set_sym &s,
-                                                       umap_vec_expr &&d);
-    vec_basic get_args() const;
-    bool is_canonical(const set_sym &vars, const umap_sym_int &degrees,
-                      const umap_vec_expr &dict);
-    std::size_t __hash__() const;
-    bool __eq__(const Basic &o) const;
-    int compare(const Basic &o) const;
-    Expression
-    eval(std::map<RCP<const Symbol>, Expression, RCPSymbolCompare> &vals) const;
-};
-
-RCP<const MultivariatePolynomial>
-add_mult_poly(const MultivariatePolynomial &a, const MultivariatePolynomial &b);
-RCP<const MultivariatePolynomial>
-neg_mult_poly(const MultivariatePolynomial &a);
-RCP<const MultivariatePolynomial>
-sub_mult_poly(const MultivariatePolynomial &a, const MultivariatePolynomial &b);
-RCP<const MultivariatePolynomial>
-mul_mult_poly(const MultivariatePolynomial &a, const MultivariatePolynomial &b);
-
-RCP<const MultivariatePolynomial> add_mult_poly(const MultivariatePolynomial &a,
-                                                const UnivariatePolynomial &b);
-RCP<const MultivariatePolynomial>
-add_mult_poly(const UnivariatePolynomial &a, const MultivariatePolynomial &b);
-RCP<const MultivariatePolynomial> sub_mult_poly(const MultivariatePolynomial &a,
-                                                const UnivariatePolynomial &b);
-RCP<const MultivariatePolynomial>
-sub_mult_poly(const UnivariatePolynomial &a, const MultivariatePolynomial &b);
-RCP<const MultivariatePolynomial> mul_mult_poly(const MultivariatePolynomial &a,
-                                                const UnivariatePolynomial &b);
-RCP<const MultivariatePolynomial>
-mul_mult_poly(const UnivariatePolynomial &a, const MultivariatePolynomial &b);
-
-RCP<const MultivariatePolynomial> add_mult_poly(const UnivariatePolynomial &a,
-                                                const UnivariatePolynomial &b);
-RCP<const MultivariatePolynomial> sub_mult_poly(const UnivariatePolynomial &a,
-                                                const UnivariatePolynomial &b);
-RCP<const MultivariatePolynomial> mul_mult_poly(const UnivariatePolynomial &a,
-                                                const UnivariatePolynomial &b);
-
-class MultivariateExprPolynomial
-{
-private:
-    RCP<const MultivariatePolynomial> poly_;
-
-public:
-    MultivariateExprPolynomial()
-    {
-    }
-    ~MultivariateExprPolynomial() SYMENGINE_NOEXCEPT
-    {
-    }
-    MultivariateExprPolynomial(const MultivariateExprPolynomial &) = default;
-    MultivariateExprPolynomial(MultivariateExprPolynomial &&other)
-        SYMENGINE_NOEXCEPT : poly_(std::move(other.poly_))
-    {
-    }
-    MultivariateExprPolynomial(RCP<const MultivariatePolynomial> p)
-        : poly_(std::move(p))
-    {
-    }
-    explicit MultivariateExprPolynomial(int i)
-    {
-        set_sym s;
-        vec_int v;
-        poly_ = MultivariatePolynomial::from_dict(s, {{v, Expression(i)}});
-    }
-
-    explicit MultivariateExprPolynomial(Expression e)
-    {
-        set_sym s;
-        vec_int v;
-        poly_ = MultivariatePolynomial::from_dict(s, {{v, e}});
-    }
-    explicit MultivariateExprPolynomial(RCP<const Symbol> sym)
-    {
-        set_sym s;
-        vec_int v;
-        poly_ = MultivariatePolynomial::from_dict(s, {{v, Expression(sym)}});
-    }
-    MultivariateExprPolynomial &operator=(const MultivariateExprPolynomial &)
-        = default;
-    MultivariateExprPolynomial &
-    operator=(MultivariateExprPolynomial &&other) SYMENGINE_NOEXCEPT
-    {
-        if (this != &other)
-            this->poly_ = std::move(other.poly_);
-        return *this;
-    }
-
-    MultivariateExprPolynomial &operator=(const Expression &other)
-    {
-        return *this = MultivariateExprPolynomial(other);
-    }
-
-    friend std::ostream &operator<<(std::ostream &os,
-                                    const MultivariateExprPolynomial &expr)
-    {
-        os << expr.poly_->__str__();
-        return os;
-    }
-
-    friend MultivariateExprPolynomial
-    operator+(const MultivariateExprPolynomial &a,
-              const MultivariateExprPolynomial &b)
-    {
-        return MultivariateExprPolynomial(add_mult_poly(*a.poly_, *b.poly_));
-    }
-
-    friend MultivariateExprPolynomial
-    operator+(const MultivariateExprPolynomial &a, const Expression &b)
-    {
-        return a + MultivariateExprPolynomial(b);
-    }
-
-    friend MultivariateExprPolynomial
-    operator+(const Expression &a, const MultivariateExprPolynomial &b)
-    {
-        return MultivariateExprPolynomial(a) + b;
-    }
-
-    MultivariateExprPolynomial &
-    operator+=(const MultivariateExprPolynomial &other)
-    {
-        poly_ = add_mult_poly(*poly_, *other.poly_);
-        return *this;
-    }
-
-    MultivariateExprPolynomial &operator+=(const Expression &other)
-    {
-        *this += MultivariateExprPolynomial(other);
-        return *this;
-    }
-
-    friend MultivariateExprPolynomial
-    operator-(const MultivariateExprPolynomial &a,
-              const MultivariateExprPolynomial &b)
-    {
-        return MultivariateExprPolynomial(sub_mult_poly(*a.poly_, *b.poly_));
-    }
-
-    friend MultivariateExprPolynomial
-    operator-(const Expression &a, const MultivariateExprPolynomial &b)
-    {
-        return MultivariateExprPolynomial(a) - b;
-    }
-
-    friend MultivariateExprPolynomial
-    operator-(const MultivariateExprPolynomial &a, const Expression &b)
-    {
-        return a - MultivariateExprPolynomial(b);
-    }
-
-    MultivariateExprPolynomial operator-() const
-    {
-        return MultivariateExprPolynomial(neg_mult_poly(*poly_));
-    }
-
-    MultivariateExprPolynomial &
-    operator-=(const MultivariateExprPolynomial &other)
-    {
-        poly_ = sub_mult_poly(*poly_, *other.poly_);
-        return *this;
-    }
-
-    MultivariateExprPolynomial &operator-=(const Expression &other)
-    {
-        *this -= MultivariateExprPolynomial(other);
-        return *this;
-    }
-
-    friend MultivariateExprPolynomial
-    operator*(const MultivariateExprPolynomial &a,
-              const MultivariateExprPolynomial &b)
-    {
-        return MultivariateExprPolynomial(
-            mul_mult_poly(*(a.poly_), *(b.poly_)));
-    }
-
-    friend MultivariateExprPolynomial
-    operator*(const MultivariateExprPolynomial &a, const Expression &b)
-    {
-        return a * MultivariateExprPolynomial(b);
-    }
-
-    friend MultivariateExprPolynomial
-    operator*(const Expression &a, const MultivariateExprPolynomial &b)
-    {
-        return MultivariateExprPolynomial(a) * b;
-    }
-
-    friend MultivariateExprPolynomial
-    operator/(const MultivariateExprPolynomial &a, const Expression &b)
-    {
-        return a * (1 / b);
-    }
-
-    MultivariateExprPolynomial operator/=(const Expression &b)
-    {
-        *this = *this / b;
-        return *this;
-    }
-
-    MultivariateExprPolynomial &
-    operator*=(const MultivariateExprPolynomial &other)
-    {
-        poly_ = mul_mult_poly(*poly_, *(other.poly_));
-        return *this;
-    }
-
-    bool operator==(const MultivariateExprPolynomial &other) const
-    {
-        return eq(*poly_, *other.poly_);
-    }
-
-    bool operator==(const Expression &other) const
-    {
-        return (*this == MultivariateExprPolynomial(other));
-    }
-
-    bool operator!=(const MultivariateExprPolynomial &other) const
-    {
-        return not(*this == other);
-    }
-
-    MultivariateExprPolynomial truncate(map_sym_uint &precs)
-    {
-        umap_vec_expr d;
-        for (auto bucket : poly_->dict_) {
-            bool accept = true;
-            auto var = poly_->vars_.begin();
-            auto prec = precs.begin();
-            unsigned int whichvar = 0;
-            while (var != poly_->vars_.end() && prec != precs.end()){
-                if ((*var)->__eq__(*prec->first)) {
-                    if (bucket.first[whichvar] > (int)(prec->second))
-                        accept = false;
-                    prec++;
-                }
-                var++;
-                whichvar++;
-            }
-            if (accept)
-                d.insert(std::pair<vec_int, Expression>(bucket.first,bucket.second));
-        }
-        return MultivariateExprPolynomial(MultivariatePolynomial::from_dict(poly_->vars_, std::move(d)));
-    }
-
-    const RCP<const MultivariatePolynomial> get_poly() const
-    {
-        return poly_;
-    }
-
-    const umap_vec_expr get_dict() const
-    {
-        return poly_->dict_;
-    }
-
-    const RCP<const Symbol> get_var() const
-    {
-        return *((poly_->vars_).begin());
-    }
-
-    const set_sym get_vars() const
-    {
-        return poly_->vars_;
-    }
-
-    const RCP<const Basic> get_basic() const
-    {
-        return poly_;
-    }
-
-}; // MultivariateExprPolynomial
-
 } // SymEngine
 
 #endif
